rename view to HtmlView and BaseElement to View
