+++ src/main/kotlin/com/narbase/kunafa/core/components/Radio.kt
@@ -16,7 +16,7 @@ import kotlin.browser.document
  * Created by islam
  * On: 10/31/17.
  */
-class Radio(parent: BaseElement? = null) : View(parent) {
+class Radio(parent: View? = null) : HtmlView(parent) {
 
     override val element: HTMLInputElement = (document.createElement("input") as HTMLInputElement).apply {
         this.type = "radio"
