+++ src/main/kotlin/com/narbase/kunafa/core/components/Svg.kt
@@ -2,18 +2,8 @@
 
 package com.narbase.kunafa.core.components
 
-import com.narbase.kunafa.core.css.*
-import com.narbase.kunafa.core.dimensions.px
-import com.narbase.kunafa.core.lifecycle.LifecycleObserver
-import com.narbase.kunafa.core.lifecycle.LifecycleOwner
-import org.w3c.dom.Element
-import org.w3c.dom.HTMLDivElement
-import org.w3c.dom.HTMLElement
-import org.w3c.dom.events.MouseEvent
 import org.w3c.dom.svg.SVGElement
 import kotlin.browser.document
-import kotlin.dom.addClass
-import kotlin.dom.removeClass
 
 /**
  * NARBASE TECHNOLOGIES CONFIDENTIAL
@@ -23,7 +13,7 @@ import kotlin.dom.removeClass
  * Created by islam
  * On: 9/30/17.
  */
-open class Svg(parent: BaseElement? = null) : BaseElement(parent) {
+open class Svg(parent: View? = null) : View(parent) {
 
     override var isViewMounted: Boolean = false
