+++ src/main/kotlin/com/narbase/kunafa/core/components/TextView.kt
@@ -11,7 +11,7 @@ package com.narbase.kunafa.core.components
  * Created by islam
  * On: 10/9/17.
  */
-open class TextView(parent: BaseElement? = null) : View(parent) {
+open class TextView(parent: View? = null) : HtmlView(parent) {
 
     var text
         get() = element.innerHTML
