+++ src/main/kotlin/com/narbase/kunafa/core/components/layout/LinearLayout.kt
@@ -2,7 +2,7 @@
 
 package com.narbase.kunafa.core.components.layout
 
-import com.narbase.kunafa.core.components.BaseElement
+import com.narbase.kunafa.core.components.HtmlView
 import com.narbase.kunafa.core.components.View
 import com.narbase.kunafa.core.components.layout.LinearLayout.Orientation.Horizontal
 import com.narbase.kunafa.core.css.*
@@ -16,9 +16,9 @@ import com.narbase.kunafa.core.css.*
  * On: 9/30/17.
  */
 open class LinearLayout(
-        parent: BaseElement?,
+        parent: View?,
         private val initialOrientation: Orientation? = Horizontal
-) : View(parent) {
+) : HtmlView(parent) {
 
     override fun configureElement() {
         super.configureElement()
