+++ src/main/kotlin/com/narbase/kunafa/core/routing/ComponentRoute.kt
@@ -2,7 +2,6 @@
 
 package com.narbase.kunafa.core.routing
 
-import com.narbase.kunafa.core.components.BaseElement
 import com.narbase.kunafa.core.components.Component
 import com.narbase.kunafa.core.components.View
 
@@ -19,8 +18,8 @@ class ComponentRoute(
         segments: List<RouteSegment>,
         val component: Component,
         parentRoute: Route?,
-        val parentView: BaseElement?,
-        private val referenceView: BaseElement,
+        val parentView: View?,
+        private val referenceView: View,
         isExact: Boolean
 ) : Route(meta, segments, parentRoute, isExact) {
     override fun onMatch(windowSegments: List<RouteSegment>) {
